+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - TransitionExpand for Collapse component
 - Disabled attribute for QuantityUpdate component
 - Pagination component
+- Price component - custom classes props
 
 ### Removed
 - -webkit- in transitions
@@ -53,6 +54,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactor styles in Select component
 - Update modal component js and styles to handle sidebar case
 - Update attributes passed to Select, Radio and Checkbox components ($attrs)
+- ProductListItem component styles for old price element
 
 ### Fixed
 - Typo in Message.selectors.json file